pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixell-sdk"
version = "0.4.15"
description = "A lightweight developer kit for packaging AI agents into portable APKG files"
readme = "README.md"
authors = [
    {name = "Pixell Core Team", email = "dev@pixell.global"}
]
license = "AGPL-3.0-only"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "watchdog>=3.0",
    "python-dotenv>=1.0",
    "tabulate>=0.9",
    "jinja2>=3.0",
    "requests>=2.28.0",
    # SDK runtime dependencies
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "pyjwt>=2.8.0",
]

[project.urls]
Homepage = "https://github.com/pixell-global/pixell-kit"
"Bug Tracker" = "https://github.com/pixell-global/pixell-kit/issues"
Documentation = "https://docs.pixell.global/pixell"
"Source Code" = "https://github.com/pixell-global/pixell-kit"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "build>=1.0",
    "twine>=4.0",
    "types-PyYAML>=6.0",
    # SDK testing dependencies
    "fakeredis>=2.20.0",
    "respx>=0.20.0",
    "pytest-timeout>=2.0",
]
signing = [
    "python-gnupg>=0.5",
]

[project.scripts]
pixell = "pixell.cli.main:cli"

[tool.setuptools]
packages = [
  "pixell",
  "pixell.cli",
  "pixell.cli.commands",
  "pixell.core",
  "pixell.models",
  "pixell.utils",
  "pixell.dev_server",
  "pixell.ui",
  "pixell.intent",
  "pixell.protocol",
  "pixell.test",
  "pixell.sdk",
  "pixell.sdk.a2a",
  "pixell.sdk.plan_mode",
  "pixell.sdk.translation",
]
include-package-data = true

[tool.setuptools.package-data]
pixell = ["templates/**/*", "protocol/schemas/*.json", "ui/schemas/*.json"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pixell --cov-report=term-missing"
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test requiring external services",
    "slow: mark test as slow-running",
]
